Use standard MODES_NET_SERVICES_NUM in faup1090
For faup1090 only the FlightAware service is enabled and we were shortening
the services table by using our own define of FAUP_NET_SERVICES_NUM and
having it set to 1.

Apps need to stick with the same number of services as defined by
MODES_NET_SERVICES_NUM in dump1090 because dump1090 support routine
modesAcceptClients loops on this constant to walk the services table.

Likewise we now define all the service ports dump1090 defines even though we
are not using them.  We have configured them as disabled.

Version bump to 1.14.