Avoid librtlsdr dependency for utilities that don't use it.
Tne only place that really needs the rtlsdr headers/libs is dump1090.c
itself. There is a dependency on rtlsdr_dev_t in the shared Modes
struct in dump1090.h, but this is a pointer type so we can just
use a dummy struct definition there to avoid the header dependency.

This lets us build faup1090 (for the piaware package) without
needing a useless librtlsdr-dev build-time dependency.