Don't throw away both CPRs, most likely only one is bad.
If we receive a new CPR and the old CPR is the one with bad data, we
will throw the data away again.  But if we are lucky and receive the
CPR that was bad, we get a good position again faster.  Also this
makes the speed check impossible on the next decode, which is bad.
Instead move the position validity closer to expiration.
That should help with wrongly accepting a bad position and failing
all speed checks afterwards.