Fix update_position being called when neither CPR changed
When we have ADS-B CPRs and get for example an MLAT synthetic CPR, the
a->cpr variables are not changed because the source is lower quality.
Yet the updatePosition function is called and lower quality CPR is used.
Fix that by only calling updatePosition if the a->cpr variables got
updated.