Introduce pos_reliable indicator as number of successful global CPRs
pos_reliable is then used to invalidate the positon if the first global
CPR to be decoded yielded a bad position.
If we have already received multiple global CPRs, a disagreeing global
CPR position means the newly decoded position is bad. This new position
is discarded while the old position stays valid.