Increase position_valid expiry time and aircraft TTL
Aircraft often drop in and out of reception. Being able to do relative
CPR and range checks even if no position was decoded for longer than 70
seconds is very useful to reception.
Due to the previously pos_reliable counter and the speed check this
should even decrease the number of wrong positions displayed while
increasing the number of aircraft relative decodes.
Especially useful for people with spotty reception or surface
positions.