Add rudimentary SBS input
Defaults to MLAT as message source.
Only the aircraft.json and VRS json output are affected by the SBS
input, outgoing or displayed messages are not displayed for the incoming
SBS messages.
At the moment, only altitude and the position are decoded from SBS, the
rest of the data is ignored.