BeastReduce output - Reduced message rate
Selectively forwards beast messages if the received data hasn't been forwarded
in the last 125 ms (BeastReduce interval).
Data not related to the physical aircraft state are only forwarded every
500 ms (4 * BeastReduce interval).
The interval can be adjusted using the commandline option.  The messages of
this output are normal beast messages and compatible with every program
able to receive beast messages.